+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 1.0.0 (unreleased)
 =====
 
+*   Add support for variations via `{{ umbrella_variations() }}`